pyproject.toml
[project]
name = "umkmall-analytics"
version = "0.1.0"
description = "Analytics platform for UMKM Mall with dbt and Lightdash"
requires-python = ">=3.11"
dependencies = [
    "dbt-core>=1.10.15",
    "dbt-postgres>=1.9.1",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "pandas>=2.1.0",
    "sqlalchemy>=2.0.0",
    "requests>=2.31.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-api-python-client>=2.100.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "ipython>=8.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ingestion"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []
